Assemble scroll toggle helper
Dialogs and drawers should lock scrolling when shown in order to prevent users from accidentally winding up somewhere else on the page after hiding a modal. Other items may require this functionality too so it is best to abstract it as a helper.
